move m, done

+++ src/interfaces/IZone.java
@@ -6,6 +6,6 @@ public interface IZone {
 	
 	public GPSSignal generateRandomGPS();
 	public String toString();
-	public int getMaxVehicles();
+	public double getMaxVehicles();
 	public void decreaseNumVehicles();
 }
